feat: reorganizando estrutura do bot
